requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
waitress==3.0.2
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
click==8.2.1
blinker==1.9.0

picamera2==0.3.31
opencv-python-headless==4.12.0.88
Pillow==11.3.0
numpy==2.2.6
simplejpeg==1.8.2
piexif==1.1.3
python-prctl==1.8.1

cryptography==46.0.1
pycryptodome==3.23.0

schedule==1.2.2
ephem==4.2
tzlocal==5.3.1
pytz==2025.2
tzdata==2025.2
python-dateutil==2.9.0.post0

pandas==2.3.2
psutil==7.0.0
lxml==6.0.1

paho-mqtt==2.1.0